+++ cmd/gokr-packer/packer.go
@@ -65,14 +65,6 @@ var gokrazyPkgs = []string{
 }
 
 func findCACerts() (string, error) {
-	// From go1.8/src/crypto/x509/root_linux.go
-	var certFiles = []string{
-		"/etc/ssl/certs/ca-certificates.crt",                // Debian/Ubuntu/Gentoo etc.
-		"/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem", // CentOS/RHEL 7
-		"/etc/pki/tls/certs/ca-bundle.crt",                  // Fedora/RHEL 6
-		"/etc/ssl/ca-bundle.pem",                            // OpenSUSE
-		"/etc/pki/tls/cacert.pem",                           // OpenELEC
-	}
 	home, err := homedir()
 	if err != nil {
 		return "", err
